requirements.txt
tensorflow
pandas 
gdown
dvc
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
joblib
types-PyYAML
scipy
Flask
Flask-Cors
langchain_groq
langchain_community
python-dotenv
streamlit 